# swagger.json
{
	"components": {
		"schemas": {
...
			"program": {
				"properties": {
					"arguments": {
						"items": {
							"$ref": "#/components/schemas/program_setting"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"webhook": {
						"properties": {
							"context": {
								"example": "some identifier included in callback to server",
								"type": "string"
							},
							"uri": {
								"example": "http://myserver.com:8002/status",
								"format": "uri",
								"type": "string"
							}
						},
						"required": [
							"uri"
						],
						"type": "object"
					}
				},
				"required": [
					"name"
				],
				"type": "object"
			},
...
			"program_setting": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"enum": [
							0,
							1,
							2,
							3,
							4,
							5,
							6,
							7,
							8
						],
						"type": "integer"
					},
					"value": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"state": {
				"properties": {
					"state": {
						"enum": [
							"Idle",
							"Emergency Stop Active",
							"Safeguard Stop Active",
							"Entity Error Active",
							"No Program",
							"Loading Program",
							"Ready",
							"Recovery",
							"Executing",
							"Pausing Execution",
							"Paused",
							"Stopping Execution",
							"Execution Error Active",
							"Joystick Active"
						],
						"type": "string"
					}
				},
				"required": [
					"state"
				],
				"type": "object"
			},
			"status": {
				"allOf": [
					{
						"$ref": "#/components/schemas/state"
					},
					{
						"properties": {
							"current_program": {
								"$ref": "#/components/schemas/program"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"battery": {
								"example": 67,
								"type": "integer"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"message": {
								"example": "Error in program (only if state is Execution Error Active)",
								"type": "string"
							}
						},
						"type": "object"
					}
				]
			},
...
		}
	},
	"info": {
		"description": "REST Interface exposed by Enabled Robotics",
		"title": "Enabled Robotics REST Interface",
		"version": "0.1-oas3"
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/status": {
			"get": {
				"description": "Get the current status of the robot",
				"operationId": "status_get",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/status"
								}
							}
						},
						"description": "Success"
					},
					"500": {
						"description": "General server error"
					},
					"503": {
						"description": "No connection to backend"
					}
				},
				"tags": [
					"Status"
				],
				"x-openapi-router-controller": "rest_backend.er_controller_v2"
			},
			"put": {
				"description": "Change the state of the robot",
				"operationId": "status_put",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"state": {
										"enum": [
											"Executing",
											"Paused",
											"Ready"
										],
										"type": "string"
									}
								},
								"required": [
									"state"
								],
								"type": "object"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/status"
								}
							}
						},
						"description": "Success. This endpoint will be deprecated in future version."
					},
					"400": {
						"description": "General user error"
					},
					"500": {
						"description": "General server error"
					},
					"503": {
						"description": "No connection to backend"
					}
				},
				"tags": [
					"Status"
				],
				"x-openapi-router-controller": "rest_backend.er_controller_v2"
			}
		},
...
	},
	"servers": [
		{
			"url": "/v2"
		}
	],
...
}